Refactors project by moving each class to its own file, and creating a src directory.
